refactor controllers with module
