+++ app/controllers/movies_controller.rb
@@ -1,5 +1,6 @@
 class MoviesController < ApplicationController
-  before_action :set_user, :remember_me
+  include ControllerHelper
+  before_action :set_user_uid, :remember_me
   def index
     @movies = MovieFacade.top20 if params[:q] == 'top rated'
     if params[:keyword]
@@ -19,15 +20,5 @@ class MoviesController < ApplicationController
     @data = MovieFacade.movie_data(params[:id])
   end
 
-  private
-  def set_user
-    @user = User.find(params[:user_id])
-  end
 
-  def remember_me
-    if !cookies[:remember_me]
-      redirect_to '/login'
-      flash[:notice] = 'Your session has expired, please log in again!'
-    end
-  end
 end
