+++ config/routes.rb
@@ -18,10 +18,4 @@ Rails.application.routes.draw do
   get 'login', to: 'users#login_form'
   post 'login', to: 'users#login'
   get 'logout', to: 'users#logout'
-  # get 'password', to: 'passwords#edit', as: 'edit_password'
-  # patch 'password', to: 'passwords#update'
-  # get 'password/reset', to: 'password_resets#new'
-  # post 'password/reset', to: 'password_resets#create'
-  # get 'password/reset/edit', to: 'password_resets#edit'
-  # patch 'password/reset/edit', to: 'password_resets#update'
 end
